requirements.txt
# selenium==3.141.0
# Appium-Python-Client==1.3.0
# django-environ==0.9.0
# Django==3.2.13
# django-jenkins==0.110.0
# urllib3==1.26.12

Django==3.2.13
selenium==4.0.0
Appium-Python-Client==2.0.0
django-environ==0.9.0
django-jenkins==0.110.0
urllib3==1.26.12
pywinauto==0.6.8
